Item content refactored, ItemData component created and ExtraFields component deleted
