+++ src/components/main/catalogues/catalogue/catalogue-item/catalogueItem.scss
@@ -14,14 +14,6 @@
     .itemContent {
         margin-left: 10px;
         width: 100%;
-        .wrapper {
-            display: flex;
-            justify-content: space-between;
-            .editButton {
-                font-size: 1.1em;
-                height: 1.1em;
-            }
-        }
     }
     .editItem {
         width: 100%;
