--- src/components/main/catalogues/catalogue/catalogue-item/extra-fields/extraFields.scss
@@ -1,7 +1,0 @@
-.extraField {
-    font-weight: 700;
-    margin-bottom: 5px;
-    span {
-        font-weight: 400;
-    }
-}
