+++ src/components/main/catalogues/catalogue/catalogue-item/item-data/itemData.scss
@@ -1,0 +1,19 @@
+.itemData {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    margin-bottom: 5px;
+    .wrapper {
+        display: flex;
+        justify-content: space-between;
+        width: 100%;
+        span {
+            font-weight: 700;
+        }
+    }
+    .editButton {
+        font-size: 1.1em;
+        height: 1.1em;
+        margin-left: 10px;
+    }
+}
